+++ hoot-rnd/src/main/cpp/hoot/rnd/io/WordCount.h
